zap-scanner/requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
python-owasp-zap-v2.4==0.0.20
requests==2.31.0
python-multipart==0.0.6
